pretrain/scripts/v3-cpt-instruct/tuning/requirements_cuda12.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.152
botocore==1.34.152
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
datasets==2.20.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
faiss-cpu==1.8.0.post1
fasteners==0.19
fasttext==0.9.3
filelock==3.15.4
Flask==3.0.3
Flask-RESTful==0.3.10
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
gdown==5.2.0
gevent==24.2.1
geventhttpclient==2.0.2
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.65.2
h5py==3.11.0
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
ijson==3.3.0
inflect==7.3.1
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lightning-utilities==0.11.6
llvmlite==0.43.0
lxml==5.2.2
Markdown==3.6
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.1
megatron_core==0.5.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nemo_aligner==0.3.1
nemo_toolkit==1.23.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numcodecs==0.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
nvidia-pytriton==0.2.5
omegaconf==2.3.0
onnx==1.16.2
OpenCC==1.1.6
packaging==24.1
pandas==2.2.2
pangu==4.0.6.1
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
protobuf==3.20.3
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.13.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-rapidjson==1.19
pytorch-lightning==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==23.2.1
rapidfuzz==3.9.5
regex==2024.7.24
requests==2.32.3
rouge_score==0.1.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.2
sacrebleu==2.4.2
sacremoses==0.1.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
sentry-sdk==2.12.0
setproctitle==1.3.3
sh==1.14.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorstore==0.1.45
text-unidecode==1.3
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.40.1
triton==3.0.0
tritonclient==2.33.0
typeguard==4.3.0
typing-inspect==0.6.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wandb==0.17.5
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.0.3
wget==3.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
youtokentome==1.0.6
zarr==2.18.2
zope.event==5.0
zope.interface==6.4.post2
